// packages/discord/src/operations/listPublicArchivedThreads.ts
export const ListPublicArchivedThreadsInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    channel_id: Schema.String.pipe(T.PathParam()),
    before: Schema.optional(Schema.String),
    limit: Schema.optional(Schema.Number),
  }).pipe(
    T.Http({
      method: "GET",
      path: "/channels/{channel_id}/threads/archived/public",
    }),
  );
export type ListPublicArchivedThreadsInput =
  typeof ListPublicArchivedThreadsInput.Type;
// ...
export type ListPublicArchivedThreadsOutput =
  typeof ListPublicArchivedThreadsOutput.Type;

// The operation
export const listPublicArchivedThreads = /*@__PURE__*/ /*#__PURE__*/ API.make(
  () => ({
    inputSchema: ListPublicArchivedThreadsInput,
    outputSchema: ListPublicArchivedThreadsOutput,
    errors: [BadRequest, Forbidden, NotFound] as const,
  }),
);